CCTV 백엔드 API DTO 구조 개선

+++ src/main/java/com/markettwin/backend/repository/PedestrianCoordinateJsonRepository.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface PedestrianCoordinateJsonRepository extends JpaRepository<PedestrianCoordinateJson, Long> {
     List<PedestrianCoordinateJson> findByClipIdAndFrameIdAndVideoId(
-            Long clipId, Integer frameId, Integer videoId);
+            Long clipId, Integer frameId, Long videoId);
 }
